Addressed bug which broke multi user authentication on same session and changed styling for dashboard. moved dashboard to its own module
