+++ static/styles/styles_dashboard.css
@@ -35,7 +35,7 @@
     margin-bottom: 20px;
   }
   
-  #settingsButton a {    
+  a {    
     text-decoration: inherit;
     color: inherit;
   }
@@ -44,7 +44,15 @@
   padding-bottom: 20px;
 }
 
-/* .api_key_item a {
-  text-decoration: inherit;
-  color: black;
-} */
+.delete_btn {
+  /* background-color: red;
+  color: white; */
+  display: block;
+  width: 100%;
+  padding: 10px;
+  background-color: red;
+  color: white;
+  border: none;
+  border-radius: 3px;
+  cursor: pointer;
+}
