+++ templates/index.html
@@ -9,6 +9,7 @@
             <div class="navbar">
                 <ul>
                     <li><a href="#">Home</a></li>
+                    <li><a href="{{links.dashboard()}}">Dashboard</a></li>
                     <li><a href="{{links.login()}}">Log In</a></li>
                     <li><a href="{{links.signup()}}">Sign Up</a></li>
                     <li><a href="#">API Doc</a></li>
