+++ ytd_web/__init__.py
@@ -9,6 +9,7 @@ from ytd_web.get_mulimedia import get_miltimedia_page
 from ytd_web.download import download_page
 from ytd_web.error import error_page
 from ytd_web.login_signup import login_signup_page
+from ytd_web.dashboard import dashboard_page
 
 endpoints: list[Blueprint] = [
     home_page, 
@@ -19,5 +20,6 @@ endpoints: list[Blueprint] = [
     get_miltimedia_page, 
     download_page, 
     error_page,
-    login_signup_page
+    login_signup_page,
+    dashboard_page
 ]
